requirements.txt
requests
celery
github3.py
Django==1.10.3
djangorestframework==3.5.3
djoser==0.5.4
requests_oauthlib==0.7.0
     